requirements.txt
django
djangorestframework
rq
redis
numpy==1.19.2
pydicom
tensorflow
wheel
barbell2light
pandas
matplotlib